Set version metadata to a semantic version number
The current version information is simply `*` and is understood by `zef`
as being effectively 0, hence another dist also by the name of
`File::Zip` is found and installed (note that [it's likely that in the
future such version strings will no longer be accepted by `zef` at
all](https://github.com/ugexe/zef/issues/553)).

Because the version number is effectively 0, this causes problems with
[`Selenium::WebDriver` not being able to install the correct `File::Zip`
dist](https://github.com/azawawi/raku-selenium-webdriver/issues/27).

This change will help rectify that issue by the version string now being
a non-wildcard character (see also #7 for another change which will help
downstream modules be able to find this version of `File::Zip`).  The
version number given here is a lower bound for the version information
and thus is an attempt to be more Raku standards-compliant.